refactor: 주문하기 관련 모델들 각각 파일을 만들어 관리하도록 변경
