+++ YABAM/Feature/Sources/Models/Menu/MenuItem.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+struct MenuItem: Hashable, Identifiable {
+    let id: UUID = UUID()
+    let name: String
+    let description: String
+    let price: Int
+    let imageName: String
+    let isRecommended: Bool
+}
