+++ YABAM/Feature/Sources/Models/Menu/MenuSection.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct MenuSection: Identifiable {
+    let id: String
+    let title: String
+    let items: [MenuItem]
+}
